culturameta/requirements.txt
Django==3.2
django-chartjs==2.1.0
django-import-export==2.0.2
django-multiselectfield==0.1.12
weasyprint==52.5
whitenoise==5.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==1.0.1
django-smart-selects==1.5.4
django-template-debug==0.3.5
djangorestframework==3.11.0
psycopg2-binary>=2.8
msgpack==0.6.2
django-cors-headers
odfpy==1.4.1
openpyxl==3.0.3
django-dynamic-raw-id
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.3.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tablib==1.1.0
Twisted==20.3.0
txaio==20.4.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
django-debug-toolbar==3.2.1
uWSGI==2.0.19.1
django-admin-interface==0.16.3